Splitting out CodeDictionary from validation_report module; created a formal change log; restructured project docs; initial version of markdown docs generator for codes.yaml codes
