+++ docs/reasoner_validator.validation_codes.rst
@@ -1,0 +1,7 @@
+Validation Codes Dictionary
+===========================
+
+.. automodule:: reasoner_validator.validation_codes
+   :members:
+   :undoc-members:
+   :show-inheritance:
