--- docs/validation_codes.md
@@ -1,1 +1,0 @@
-# Validation Message Code Dictionary
